requirements/production.txt
-r base.txt

django-toolbelt==0.0.1
Pillow==2.3.0
boto==2.19.0
django-storages==1.1.8